chore: add support for swipe down gesture
